requirements.txt
pillow~=10.4.0
termcolor~=2.4.0
requests~=2.32.3
torch~=2.4.1
uvicorn~=0.31.1
fastapi~=0.115.0
fastchat~=0.1.0
transformers~=4.45.2
albumentations~=1.4.18
tqdm~=4.66.5
numpy~=2.1.2
opencv-python~=4.10.0.84
peft~=0.13.2
pandas~=2.2.3